requirements.txt
alembic==1.11.1
bleach==6.0.0
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
dominate==2.8.0
ecdsa==0.18.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
packaging==23.1
pyasn1==0.5.0
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.16
typing-extensions==4.6.3
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
